Make the UI be parameterised by the calculus
Turn ProofTree_UI into a functor that takes a description of the
calculus, and a description of partial proofs, and turns it into a
proof-tree building UI.

Not entirely happy with how this has turned out. There is too much
duplication between the definition of the rules for the SYSTEM
signature, and the definition of the partials for the PARTIALS
signature.